requirements.txt
fastapi
uvicorn[standard]
python-dotenv
nltk
faster-whisper
soundfile
numpy
google-generativeai
pydub
llama-cpp-python
onnxruntime
ffmpeg-python
transformers
webrtcvad-wheels
scipy
openai
optimum[onnxruntime]
bitsandbytes
psutil
kittentts
num2words
kokoro-tts
httpx
setuptools<80